Clean up s3 lambda and start unit testing
